backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.115.0
uvicorn[standard]==0.34.0
python-multipart==0.0.18
pydantic==2.5.0
slowapi==0.1.9

# Sprint 1: Document parsing
python-pptx==0.6.21
PyMuPDF==1.23.8
pymupdf==1.23.8
pdf2image==1.17.0
Pillow>=10.0.0
# OpenCV headless для серверного окружения (без GUI зависимостей)
# Используется для оптимизированной генерации видео кадров (в ~5 раз быстрее PIL)
opencv-python-headless>=4.8.0
numpy>=1.24.0,<2.0.0
pytesseract==0.3.10
easyocr==1.7.0

# Sprint 2: AI and TTS
openai==1.54.0
anthropic==0.7.8
elevenlabs==0.2.26
azure-cognitiveservices-speech==1.34.0

# Silero TTS (Local, Free)
torch>=2.0.0
torchaudio>=2.0.0
omegaconf>=2.3.0

# ✅ Whisper for accurate word-level timing from TTS audio
openai-whisper>=20231117

# Sprint 3: Video export and queues
celery==5.3.4
redis==5.0.1
ffmpeg-python==0.2.0
moviepy==1.0.3
pydub==0.25.1
boto3==1.34.0
botocore==1.34.0

# Google Cloud Services
# ✅ Обновлено до 2.32.0 для поддержки Gemini 2.5 Flash TTS с русским языком
google-cloud-texttospeech==2.32.0
google-cloud-documentai==2.20.1
google-cloud-aiplatform==1.38.1
google-cloud-storage==2.10.0
google-cloud-vision==3.4.4

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
prometheus_client==0.20.0
tenacity==8.2.3
python-magic==0.4.27
cachetools==5.3.3
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.0
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
redis==5.0.1
cryptography>=41.0.0
user-agents==2.2.0

# Error tracking and monitoring
sentry-sdk[fastapi]==1.45.1

# Payment processing
stripe==7.0.0

# Translation and language detection
langdetect==1.0.9
google-cloud-translate>=3.16.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
google-generativeai>=0.8.0
